move tab stack to tab index
